+++ tests/testthat/test-map_procrustes.R
@@ -143,17 +143,18 @@ test_that("Procrustes a map to itself", {
 })
 
 
-test_that("Procrustes a map to itself", {
+test_that("Procrustes a map to one with no matching viruses", {
 
   map1a <- map1
   agNames(map1a) <- paste(agNames(map1a), "alt")
   srNames(map1a) <- paste(srNames(map1a), "alt")
   expect_error(
     procrustesData(map1, map1a),
-    "No matching points"
+    "Not enough matching points \\(0\\)"
   )
 
 })
+
 
 test_that("Realign to a transformed version", {