+++ NEWS.md
@@ -27,3 +27,4 @@
 * Fix error in viewer code when accounting for antigen reactivity adjustments in stress calculations
 * Add option in `plot.acmap()` to draw stress lines
 * Add xlim and ylim options to `view.acmap()` to control how maps are displayed
+* Add check to procrustes that there are sufficient matching points